+++ extensions/api/data-management/contractnegotiation/src/main/java/org/eclipse/dataspaceconnector/api/datamanagement/contractnegotiation/ContractNegotiationApiExtension.java
@@ -8,27 +8,63 @@
  * SPDX-License-Identifier: Apache-2.0
  *
  * Contributors:
- *   ZF Friedrichshafen AG - Initial API and Implementation
+ *    ZF Friedrichshafen AG - Initial API and Implementation
+ *    Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
  */
 
 package org.eclipse.dataspaceconnector.api.datamanagement.contractnegotiation;
 
 import org.eclipse.dataspaceconnector.api.datamanagement.configuration.DataManagementApiConfiguration;
+import org.eclipse.dataspaceconnector.api.datamanagement.contractnegotiation.service.ContractNegotiationServiceImpl;
+import org.eclipse.dataspaceconnector.api.datamanagement.contractnegotiation.transform.ContractAgreementToContractAgreementDtoTransformer;
+import org.eclipse.dataspaceconnector.api.datamanagement.contractnegotiation.transform.ContractNegotiationToContractNegotiationDtoTransformer;
+import org.eclipse.dataspaceconnector.api.datamanagement.contractnegotiation.transform.NegotiationInitiateRequestDtoToDataRequestTransformer;
+import org.eclipse.dataspaceconnector.api.transformer.DtoTransformerRegistry;
 import org.eclipse.dataspaceconnector.spi.WebService;
+import org.eclipse.dataspaceconnector.spi.contract.negotiation.ConsumerContractNegotiationManager;
+import org.eclipse.dataspaceconnector.spi.contract.negotiation.store.ContractNegotiationStore;
 import org.eclipse.dataspaceconnector.spi.system.Inject;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
+import org.eclipse.dataspaceconnector.spi.transaction.NoopTransactionContext;
+import org.eclipse.dataspaceconnector.spi.transaction.TransactionContext;
+
+import static java.util.Optional.ofNullable;
 
 public class ContractNegotiationApiExtension implements ServiceExtension {
-    @Inject
-    private WebService webService;
 
     @Inject
-    private DataManagementApiConfiguration config;
+    WebService webService;
+
+    @Inject
+    DataManagementApiConfiguration config;
+
+    @Inject
+    DtoTransformerRegistry transformerRegistry;
+
+    @Inject
+    ContractNegotiationStore store;
+
+    @Inject
+    ConsumerContractNegotiationManager manager;
+
+    @Inject(required = false)
+    TransactionContext transactionContext;
 
     @Override
     public void initialize(ServiceExtensionContext context) {
+        transformerRegistry.register(new ContractNegotiationToContractNegotiationDtoTransformer());
+        transformerRegistry.register(new ContractAgreementToContractAgreementDtoTransformer());
+        transformerRegistry.register(new NegotiationInitiateRequestDtoToDataRequestTransformer());
+
         var monitor = context.getMonitor();
-        webService.registerResource(config.getContextAlias(), new ContractNegotiationApiController(monitor));
+        var transactionContextImpl = ofNullable(transactionContext)
+                .orElseGet(() -> {
+                    monitor.warning("No TransactionContext registered, a no-op implementation will be used, not suitable for production environments");
+                    return new NoopTransactionContext();
+                });
+
+        var service = new ContractNegotiationServiceImpl(store, manager, transactionContextImpl);
+        webService.registerResource(config.getContextAlias(), new ContractNegotiationApiController(monitor, service, transformerRegistry));
     }
 }